moved get dr as dict to service, new dr route in rest route. new template for lightcurve app
